render each card item in its own component
